move product filters panel
