+++ src/app/api/data/http/impl/products.http-data-api.service.ts
@@ -7,7 +7,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Product } from 'src/app/models/entities/Product';
-import { ProductFilters } from 'src/app/shared/product-filters-panel/product-filters-panel.component';
+import { ProductFilters } from 'src/app/shared/components/product-filters-panel/product-filters-panel.component';
 import { EntityDataApiIService } from '../../entity-data-api.iservice';
 import { HttpDataApiService } from '../http-data-api.aservice';
