+++ src/app/shared/components/product-filters-panel/product-filters-panel.component.ts
