+++ src/app/shared/components/product-filters-panel/product-filters-panel.service.spec.ts
